Main menu flows work as well as pause menu with pausing, and unpausing.
Press M to pause and extract (will be esc in game but editor uses esc).
Menus should all flow together, only one technically missing is the mission complete menu which is there and attached to the main menu, but hasnt been set up to trigger in-game when the mission is completed.